Inline KeyValueStorageTest into DBKeyValueStorageTest
